labels displaing in variable rows and group select but not in filter labels and data listing headers
