+++ src/codebook/controls/groups/update.js
@@ -9,11 +9,17 @@ export function update(codebook) {
     groupSelect = groupControl.select("select"),
     columns = Object.keys(codebook.data.raw[0]),
     groupLevels = d3merge([
-      ["None"],
-      codebook.config.groups.map(m => m.value_col)
+      [{value_col: "None", label: 'None'}],
+      codebook.config.groups
+        .map(m => {
+            return {
+                value_col: m.value_col,
+                label: codebook.data.summary
+                    .filter(variable => variable.value_col === m.value_col)[0]
+                    .label}; })
     ]),
-    groupOptions = groupSelect.selectAll("option").data(groupLevels, d => d);
-  groupOptions.enter().append("option").text(d => d);
+    groupOptions = groupSelect.selectAll("option").data(groupLevels, d => d.value_col);
+  groupOptions.enter().append("option").property('label', d => d.label).text(d => d.value_col);
   groupOptions.exit().remove();
   groupOptions.sort((a, b) => columns.indexOf(a) - columns.indexOf(b));
   groupSelect.on("change", function() {
