+++ src/codebook/data/summarize/index.js
@@ -1,0 +1,8 @@
+import determineType from './determineType';
+import categorical from './categorical';
+import continuous from './continuous';
+
+export default
+    {determineType: determineType
+    ,categorical: categorical
+    ,continuous: continuous};
