+++ src/codebook/defaultSettings.js
@@ -1,6 +1,7 @@
 const defaultSettings = {
   filters: [],
   groups: [],
+  variableLabels: [],
   autogroups: 5, //automatically include categorical vars with 2-5 levels in the groups dropdown
   autofilter: 10, //automatically make filters for categorical variables with 2-10 levels
   autobins: true,
