+++ src/codebook/init.js
@@ -26,6 +26,7 @@ export function init(data) {
   //draw controls
   this.util.makeAutomaticFilters(this);
   this.util.makeAutomaticGroups(this);
+    console.log(this.data.summary);
   this.controls.init(this);
 
   //initialize nav
