+++ src/codebook/summaryTable/renderRow/makeTitle.js
@@ -11,7 +11,7 @@ export default function makeTitle(d) {
   var valuesList = wrap.append("ul").attr("class", "value-list");
 
   //Title and type
-  titleDiv.append("div").attr("class", "name").html(d => d.value_col);
+  titleDiv.append("div").attr("class", "name").html(d => d.label);
   titleDiv.append("div").attr("class", "type").html(d => d.type);
 
   //make a list of values
